requirements/base.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy[asyncio]
psycopg[binary]  # For async PostgreSQL connection
python-dotenv
alembic
aiosqlite 
crewai
crewai-tools
langchain-openai  # Or another provider like langchain-anthropic
python-decouple # A good library for handling API keys
arq
redis